revert xgboost-predictor-performant-op
