+++ RELEASE_NOTES.md
@@ -7,7 +7,6 @@
 - Scikit-learn support for Multinomial Logistic Regression
 - Support for min/max values other than defaults (i.e. 0.0 and 1.0) in MinMaxScalerModel
 - Support for custom transformers (StringMap, MathUnary) in Pyspark
-- Support for a performant implementation of the XGboost runtime (XGboost Predictor)
 
 ### Bug Fixes
 - Fix XGBoost sparse vector support
