+++ project/Common.scala
@@ -22,7 +22,6 @@ object Common {
     fork in Test := true,
     javaOptions in test += sys.env.getOrElse("JVM_OPTS", ""),
     resolvers += Resolver.mavenLocal,
-    resolvers += Resolver.jcenterRepo,
     resolvers ++= {
       // Only add Sonatype Snapshots if this version itself is a snapshot version
       if(isSnapshot.value) {
